Extracts RXMagnitude into its own header.
